Refactor tests and datetime parsing
